Fixes #19: Material doesn't return days if given as input
Also cleaning up the objects package
